refactor(desktop): split oversized feature components
