+++ apps/desktop/src/features/session/components/NewSessionView/index.tsx
@@ -1,5 +1,5 @@
 import { invoke } from '@tauri-apps/api/core';
-import { useEffect, useRef, useState, type ReactNode } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { Button, Divider, Input, ScrollFade, Skeleton, Textarea, cn } from '@goodboy/ui';
 import { AlertTriangle, GitBranch, Paperclip, Target, Wand2 } from 'lucide-react';
 import type { ProviderId, SessionId, WorkspaceId } from '@goodboy/types';
@@ -27,6 +27,8 @@ import { isValidBranchSlug as validateBranchSlug } from '../../../../shared/util
 import { sanitizeBranchPrefix } from '../../../../shared/utils/sanitizeBranchPrefix';
 import { sanitizeBranchSlug as sanitizeBranchSlugValue } from '../../../../shared/utils/sanitizeBranchSlug';
 import { slugifyBranch } from '../../../../shared/utils/slugifyBranch';
+import { BranchModeToggle } from './BranchModeToggle';
+import { Section } from './Section';
 
 type Props = {
   onClose: () => void;
@@ -575,89 +577,3 @@ export const NewSessionView = ({ onClose, workspaceId, onOpenSettings }: Props) 
     </div>
   );
 };
-
-type Tone = 'primary' | 'success';
-
-const TONE_BG: Record<Tone, string> = {
-  primary: 'bg-primary/10',
-  success: 'bg-success/10',
-};
-
-function Section({
-  icon,
-  tone,
-  title,
-  subtitle,
-  children,
-}: {
-  icon: ReactNode;
-  tone: Tone;
-  title: string;
-  subtitle: string;
-  children: ReactNode;
-}) {
-  return (
-    <section className="flex flex-col gap-2">
-      <header className="flex items-start gap-2">
-        <span
-          className={cn(
-            'mt-0.5 flex h-7 w-7 shrink-0 items-center justify-center rounded-md',
-            TONE_BG[tone],
-          )}
-        >
-          {icon}
-        </span>
-        <div className="flex flex-col gap-0.5">
-          <h3 className="text-sm font-semibold text-foreground">{title}</h3>
-          <p className="text-2xs leading-relaxed text-muted-foreground">{subtitle}</p>
-        </div>
-      </header>
-      {children}
-    </section>
-  );
-}
-
-function BranchModeToggle({
-  mode,
-  onChange,
-  disabled,
-}: {
-  mode: 'new' | 'existing';
-  onChange: (next: 'new' | 'existing') => void;
-  disabled: boolean;
-}) {
-  const modes: ReadonlyArray<{ id: 'new' | 'existing'; label: string }> = [
-    { id: 'new', label: 'New' },
-    { id: 'existing', label: 'Existing' },
-  ];
-  return (
-    <div
-      role="tablist"
-      aria-label="branch source"
-      className="inline-flex shrink-0 rounded border border-border bg-background p-0.5"
-    >
-      {modes.map((m) => {
-        const active = mode === m.id;
-        return (
-          <button
-            key={m.id}
-            type="button"
-            role="tab"
-            aria-selected={active}
-            disabled={disabled}
-            onClick={() => onChange(m.id)}
-            className={cn(
-              'rounded px-2 py-0.5 text-2xs font-medium motion-safe:transition-colors',
-              active
-                ? 'bg-muted text-foreground shadow-sm'
-                : 'text-muted-foreground hover:text-foreground',
-              disabled && 'cursor-not-allowed opacity-50',
-            )}
-          >
-            {m.label}
-          </button>
-        );
-      })}
-    </div>
-  );
-}
