+++ apps/desktop/src/features/session/components/WorkflowBuilderView/parts/LaunchToggleRow.tsx
@@ -1,0 +1,30 @@
+import { ToggleSwitch } from '../../../../../shared/components/ToggleSwitch';
+
+type Props = {
+  readonly title: string;
+  readonly description: string;
+  readonly checked: boolean;
+  readonly onChange: (v: boolean) => void;
+  readonly disabled: boolean;
+  readonly beta?: boolean;
+};
+
+export const LaunchToggleRow = ({
+  title,
+  description,
+  checked,
+  onChange,
+  disabled,
+  beta,
+}: Props) => (
+  <div className="flex items-center justify-between gap-3 px-3 py-2.5">
+    <p className="min-w-0 text-2xs leading-relaxed text-muted-foreground/60">{description}</p>
+    <ToggleSwitch
+      label={title}
+      beta={beta}
+      checked={checked}
+      onChange={onChange}
+      disabled={disabled}
+    />
+  </div>
+);
