+++ apps/desktop/src/features/settings/components/GuideStudio/parts/findScrollParent.ts
@@ -1,0 +1,15 @@
+type Params = {
+  readonly element: HTMLElement;
+};
+
+export const findScrollParent = ({ element }: Params): HTMLElement | null => {
+  let parent = element.parentElement;
+  while (parent) {
+    const overflowY = getComputedStyle(parent).overflowY;
+    if (overflowY === 'auto' || overflowY === 'scroll') {
+      return parent;
+    }
+    parent = parent.parentElement;
+  }
+  return null;
+};
